[CDRIVER-5859] Enable sign-compare warnings globally, and fix them all (#1856)
* Add a statement macro that can assert SIGABRT

The `mlib_assert_aborts` macro checks that the following statement
terminates the process with SIGABRT. This relies on `fork()`, so it only
works on Unix systems. On Win32 it is a no-op.

* New test assertion macro
* Integer range looping macro.

`mlib_foreach_irange` and `mlib_foreach_urange` provide
concise looping over an integral range.

* Looping over arrays
* Replace all sign-compare loops with macro loops.

This change swaps any `for (...)` loops that trigger
`-Wsign-compare` with `mlib/loop` macros. This also simplifies
redundant code around looping over array elements.

* Integer encoding utilities.

This adds function for reading little-endian i32, u32, i64, and u64 from
pointers to memory. This allows us to decode integers in a single line
instead of doing a declare+memcpy+byteswap that clutters the code and
prevents us from using `const` and correct signedness. Instead, we can
declare and initialize integers of the exact size and sign that we want
in a single line.

* Use integer decoding functions throughout the codebase
* Enable sign-compare warnings globally, and fix every occurrence

+++ src/libmongoc/tests/TestSuite.h
@@ -183,8 +183,8 @@ _test_error (const char *format, ...) BSON_GNUC_PRINTF (1, 2);
             int fd2 = bson_open ("failure.expected.bson", O_RDWR | O_CREAT, 0640);                              \
             ASSERT (fd1 != -1);                                                                                 \
             ASSERT (fd2 != -1);                                                                                 \
-            ASSERT ((bson)->len == bson_write (fd1, bson_data, (bson)->len));                                   \
-            ASSERT ((expected)->len == bson_write (fd2, expected_data, (expected)->len));                       \
+            ASSERT ((int) (bson)->len == (int) bson_write (fd1, bson_data, (bson)->len));                       \
+            ASSERT ((int) (expected)->len == (int) bson_write (fd2, expected_data, (expected)->len));           \
             bson_close (fd1);                                                                                   \
             bson_close (fd2);                                                                                   \
          }                                                                                                      \
